moved environment to separate file
